Moved getEventTitle logic to geoserve places service. Stop if no lat/lng. Handle exceptions. Shorter timeouts.
