+++ src/main/java/gov/usgs/earthquake/geoserve/GeoservePlacesService.java
@@ -8,6 +8,7 @@ import java.net.URL;
 import java.net.URLEncoder;
 
 import javax.json.Json;
+import javax.json.JsonArray;
 import javax.json.JsonObject;
 import javax.json.JsonReader;
 
@@ -41,12 +42,58 @@ public class GeoservePlacesService {
     final URL url = new URL(String.format("%s?type=event&latitude=%s&longitude=%s", this.endpointUrl,
         URLEncoder.encode(latitude.toString(), "UTF-8"), URLEncoder.encode(longitude.toString(), "UTF-8")));
 
-    try (InputStream in = StreamUtils.getInputStream(url)) {
+    try (InputStream in = StreamUtils.getURLInputStream(url, 250, 250)) {
       JsonReader reader = Json.createReader(in);
       JsonObject json = reader.readObject();
       reader.close();
       return json.getJsonObject("event");
     }
+  }
+
+  public String getEventTitle(BigDecimal latitude, BigDecimal longitude) throws IOException, MalformedURLException {
+    JsonObject places = this.getEventPlaces(latitude, longitude);
+    JsonArray features = places.getJsonArray("features");
+    JsonObject feature = features.get(0).asJsonObject();
+    JsonObject properties = feature.getJsonObject("properties");
+
+    String name = properties.getString("name");
+    String country = properties.getString("country_code").toLowerCase();
+    String admin = properties.getString("country_name");
+    int distance = properties.getInt("distance");
+    double azimuth = properties.getJsonNumber("azimuth").doubleValue();
+    String direction = azimuthToDirection(azimuth);
+
+    if ("us".equals(country)) {
+      admin = properties.getString("admin1_name");
+    }
+
+    return String.format("%d km %s of %s, %s", distance, direction, name, admin);
+
+  }
+
+  /**
+   * Converts a decimal degree azimuth to a canonical compass direction
+   *
+   * @param {Number} azimuth The degrees azimuth to be converted
+   *
+   * @return {String} The canonical compass direction for the given input azimuth
+   */
+  public String azimuthToDirection(double azimuth) {
+    double fullwind = 22.5;
+    String[] directions = { "N", "NNE", "NE", "ENE", "E", "ESE", "SE", "SSE", "S", "SSW", "SW", "WSW", "W", "WNW", "NW",
+        "NNW", "N" };
+
+    // Invert azimuth for proper directivity
+    // Maybe not needed in the future.
+    azimuth += 180.0;
+
+    // adjust azimuth if negative
+    azimuth = Math.abs(azimuth);
+    while (azimuth < 0.0) {
+      azimuth = azimuth + 360.0;
+    }
+
+    return directions[(int) Math.round((azimuth % 360.0) / fullwind)];
   }
 
   // TODO as needed, implement full GeoServe places API options
